+++ GDX/Runtime/HideFromDocFXAttribute.cs
@@ -1,0 +1,24 @@
+﻿// dotBunny licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+
+namespace GDX
+{
+    /// <summary>
+    ///     Hide the attributed item from DocFX documentation generation.
+    /// </summary>
+    /// <example>
+    ///     <para>
+    ///         For this to work a custom entry in a projects <c>filterConfig.yml</c> is necessary to define the exclusion.
+    ///     </para>
+    ///     <code>
+    ///     - exclude:
+    ///         hasAttribute:
+    ///         uid: GDX.HideFromDocFXAttribute
+    ///     </code>
+    /// </example>
+    public sealed class HideFromDocFXAttribute : Attribute
+    {
+    }
+}