+++ GDX/Runtime/InspectorLabelAttribute.cs
@@ -1,10 +1,13 @@
 ﻿// dotBunny licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+#if UNITY_EDITOR
+
 namespace GDX
 {
     using UnityEngine;
 
+    [HideInInspector]
     /// <summary>
     ///     Override the label content of the scripting variable.
     /// </summary>
@@ -21,7 +24,9 @@ namespace GDX
         /// <param name="label">The labels <see cref="GUIContent.text"/>.</param>
         public InspectorLabelAttribute(string label)
         {
-            Label = label ;
+            Label = label;
         }
     }
 }
+
+#endif