+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Optimized `GDX.GDXConfig` loading at runtime.
+- Wrap `GDX.InspectorLabelAttribute` in `UNITY_EDITOR` define requirement.
 ### Removed
 - Removed centralized `GDX.Strings`, moving ownership of data to the actual primary consumer.
 - Removed reference to unit test locations from files.